backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.5
celery==5.2.7
redis==4.5.1
